requirements.txt
alembic==1.17.2
amqp==5.3.1
beautifulsoup4==4.14.3
billiard==4.2.4
blinker==1.9.0
celery==5.6.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
exceptiongroup==1.3.1
Flask==3.1.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kombu==5.6.1
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.4.0
packaging==25.0
pandas==2.3.3
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.0
requests==2.32.5
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
soupsieve==2.8.1
SQLAlchemy==2.0.45
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.4
